requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
python-dotenv>=1.0

# OpenAI
openai>=1.40
openai-agents>=0.2.10

# Documents
python-docx>=1.1
pdfplumber>=0.11

# Optional but useful for token accounting
tiktoken>=0.7
